+++ CHANGELOG.rst
@@ -3,6 +3,7 @@
 
 - Restrict dependencies to those of matplotlib 1.5 (still Python 3 only).
 - Automatic positioning and alignment of annotation text.
+- `Patch`\es now pick on their borders, not their interior.
 
 
 0.1